Mangal

Mangal

Login

Mangal.login — Function.

This function will store the token in the MANGAL_BEARER_TOKEN environmental variable.

source

Verbosity

Mangal.verbose — Function.

verbose()

This function will switch the package to verbose mode.

source

verbose(vrb::Bool)

This function will switch the package to verbose mode, or silence it.

source

Mangal.isverbose — Function.

isverbose()

This function will return a Boolean for the current package verbosity.

source

List of types

Mangal.MangalDataset — Type.

A MangalDataset identifies a collection of networks, possibly containing a single element. A dataset is identified by its id or name (both of which are unique).

source

Mangal.MangalNetwork — Type.

Wrapper for species interactions

source

Mangal.MangalReferenceTaxon — Type.

Reference taxon (unique identifier of network nodes)

source

Mangal.MangalNode — Type.

Node in a network

The taxon field is a MangalReferenceTaxon object, so that one can, for example, query the TSN identifier of a node through object.taxon.tsn.

This approach has been chosen because (i) names of nodes in networks can be non unique and (ii) nodes within the same networks can refer to various taxonomic levels. As an example, if a network has four distinct nodes identified as Ascariasis sp., they will represent four nodes in the networks, but map onto the same MangalReferenceTaxon (representing the entire Ascariasis genus). This approach provides a seemless representation of the same taxon across different networks, but also of the same taxon within networks.

source

Methods

Queries

All queries are passed as vectors of pairs. For example, filtering interactions that are of the mutualist type can be done with [Pair("type", "mutualism")].

For datasets

Mangal.datasets — Function.

datasets()

When called with no arguments, this function will return a list of the most recent datasets. The results will be returned as a vector of MangalDataset object.

source

datasets(q::Vector{Pair{String,T}}) where {T <: Any}

Will return the most recent datasets that match a given query. The results will be returned as a vector of MangalDataset object.

source

Mangal.dataset — Function.

dataset(name::AbstractString)

Return a single dataset by its name.

source

dataset(id::Int64)

Return a single dataset by its unique numerical ID.

source

Networks

networks
network

Internals

Formatters

Mangal.format_dataset_response
Mangal.format_network_response
Mangal.format_node_response
Mangal.format_backbone_response

Other functions

Mangal.generate_base_header
Mangal.generate_request_query
Mangal.search_objects_by_query