Internal functions

Login

Mangal.login — Function.

This function will store the token in the MANGAL_BEARER_TOKEN environmental variable.

source

Verbosity

Mangal.verbose — Function.

verbose()

This function will switch the package to verbose mode.

source

verbose(vrb::Bool)

This function will switch the package to verbose mode, or silence it.

source

Mangal.isverbose — Function.

isverbose()

This function will return a Boolean for the current package verbosity.

source

Formatters

Mangal.format_dataset_response
Mangal.format_network_response
Mangal.format_node_response
Mangal.format_backbone_response

Other functions

Mangal.generate_base_header
Mangal.generate_request_query
Mangal.search_objects_by_query

Caching

Mangal.cache — Function.

Internally, the Mangal package uses a cache to store some objects that are likely to be queried more than once. These are MangalNode and MangalReferenceTaxon, which are called in a nested way during the querying of e.g. Interactions. This is not a fancy mechanism, and it only works when calling the nodes or backbones by their id (which is what the resources-hungry functions do internally anyways).

source