Getting data

Queries

All queries are passed as vectors of pairs. For example, filtering interactions that are of the mutualist type can be done with [Pair("type", "mutualism")].

A convenient shortcut

Mangal.getallof — Function.

gimme(::Type{MangalNode}, n::MangalNetwork)

Returns all MangalNode objects contained into the MangalNetwork object passed as its second argument.

source

gimme(::Type{MangalNetwork}, n::MangalDataset)

Returns all MangalNetwork objects contained into the MangalDataset object passed as its second argument.

source

For datasets

Mangal.datasets — Function.

datasets()

When called with no arguments, this function will return a list of the most recent datasets. The results will be returned as a vector of MangalDataset object.

source

datasets(q::Vector{Pair{String,T}}) where {T <: Any}

Will return the most recent datasets that match a given query. The results will be returned as a vector of MangalDataset object.

source

Mangal.dataset — Function.

dataset(name::AbstractString)

Return a single dataset by its name.

source

dataset(id::Int64)

Return a single dataset by its unique numerical ID.

source

Networks

networks
network

Interactions

interactions
interaction

Nodes

nodes
node

Reference taxon

Mangal.backbone — Function.

backbone()

source