Base.count — Function

count(::Type{MangalReferenceTaxon}, query::Pair...)

Returns the number of MangalReferenceTaxon objects that match a query.

count(::Type{MangalNode}, query::Pair...)

Returns the number of MangalNode objects that match a query.

count(::Type{MangalNetwork}, query::Pair...)

Returns the number of MangalNetwork objects that match a query.

count(::Type{MangalDataset}, query::Pair...)

Returns the number of MangalDataset objects that match a query.

count(::Type{MangalReference}, query::Pair...)

Returns the number of MangalReference objects that match a query.

count(::Type{MangalInteraction}, query::Pair...)

Returns the number of MangalInteraction objects that match a query.

count(::Type{MangalAttribute}, query::Pair...)

Returns the number of MangalAttribute objects that match a query.

count(::Type{MangalUser}, query::Pair...)

Returns the number of MangalUser objects that match a query.

count(::Type{MangalNetwork}, d::MangalDataset, query::Pair...)

Return the number of networks that belong to a dataset, according to an optional series of queries.

count(::Type{MangalInteraction}, n::MangalNetwork, query::Pair...)

Return the number of interactions that belong to a network, according to an optional series of queries.

count(::Type{MangalNode}, n::MangalNetwork, query::Pair...)

Return the number of nodes that belong to a network, according to an optional series of queries.

count(::Type{MangalNode}, n::MangalReferenceTaxon, query::Pair...)

Return the number of nodes that are instances of a reference taxon, according to an optional series of queries.