Getting data

The recommended way to get data is through Mangal.jl integration with the EcologicalNetworks.jl package for networks analysis.

There is no obvious way to know in advance if a network is bipartite or not. For this reason, this wrapper only returns unipartite objects. They can be converted into bipartite networks using the convert methods in EcologicalNetworks.jl.

Base.convert — Function.

convert(::Type{UnipartiteNetwork}, n::MangalNetwork; resolution::Type=MangalNode)

Returns a UnipartiteNetwork object representation of the MangalNetwork passed as its first argument. The optional keyword resolution (can be MangalNode or MangalReferenceTaxon) is determined to use which level of aggregation should be used. The default (MangalNode) is raw data, and MangalReferenceTaxon is the cleaned version.

source