Data types

Core types

Mangal.MangalDataset — Type.

A MangalDataset identifies a collection of networks, possibly containing a single element. A dataset is identified by its id or name (both of which are unique).

id (Int64): a unique identifier for the dataset.

name (AbstractString): a unique name describing the dataset.

public (Bool): indicates whether the dataset details are available to others than its owner.

date (DateTime): date and time at which the dataset was assembled. This can refer to the sampling time of networks, or to the date at which the dataset was finalized.

created (DateTime): date and time at which the dataset was created in the database.

updated (DateTime): date and time at which the dataset was last modified in the database. For most datasets, this will be equal to created.

reference (Union{Int64,Nothing}) (optional): a reference to the id of the MangalReference, or nothing if there is no associated reference for this dataset.

user (Int64): id of the user who added the dataset to the database. This is not necessarily the author of the dataset, see reference (and the same field in the MangalNetwork) to get the actual authorship.

description (AbstractString): a free-form description of the dataset.

source

Mangal.MangalNetwork — Type.

Wrapper for species interactions

source

Mangal.MangalInteraction — Type.

Interaction

source

Taxonomy types

Mangal.MangalNode — Type.

Node in a network

The taxon field is a MangalReferenceTaxon object, so that one can, for example, query the TSN identifier of a node through object.taxon.tsn.

This approach has been chosen because (i) names of nodes in networks can be non unique and (ii) nodes within the same networks can refer to various taxonomic levels. As an example, if a network has four distinct nodes identified as Ascariasis sp., they will represent four nodes in the networks, but map onto the same MangalReferenceTaxon (representing the entire Ascariasis genus). This approach provides a seemless representation of the same taxon across different networks, but also of the same taxon within networks.

source

Mangal.MangalReferenceTaxon — Type.

Reference taxon (unique identifier of network nodes)

source

Additional information

MangalTrait
MangalAttribute

Metadata types

MangalUser
MangalReference