Internal functions

Login

Login

Mangal relies on ORCID for authentication and login. As long as you have an ORCID profile, you can login.

Mangal.login — Function.

login(token::AbstractString)

This function will store the token in the MANGAL_BEARER_TOKEN environmental variable. To get the your token, please use login with no argument.

source

login()

Read the bearer token from the MANGAL_BEARER_TOKEN environment variable. If not found, displays a login message with a login URL.

source

Formatters

Missing docstring.

Missing docstring for Mangal.format_mangal_response. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Mangal.format_mangal_coordinates. Check Documenter's build log for details.

Other functions

Mangal.generate_base_header — Function.

generate_base_header()

If a bearer token is present, this function will add it to the header.

source

Missing docstring.

Missing docstring for Mangal.generate_request_query. Check Documenter's build log for details.

Mangal.search_objects_by_query — Function.

search_objects_by_query(endpoint::AbstractString, formatter::Function, query::Pair...)

In all cases, it is assumed that the functions will be wrapepd in calls to query objects until no further objects are found.

source

Caching

Mangal.cache — Function.

Internally, the Mangal package uses a cache to store some objects that are likely to be queried more than once. These are MangalNode and MangalReferenceTaxon, which are called in a nested way during the querying of e.g. Interactions. This is not a fancy mechanism, and it only works when calling the nodes or backbones by their id (which is what the resources-hungry functions do internally anyways).

source

Full data retrieval

Missing docstring.

Missing docstring for Mangal.get_all_nodes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Mangal.get_all_interactions. Check Documenter's build log for details.