Methods for data retrieval

A note on paging and package design

Paging matters!

The server returns (by default) 50 objects for a given query, and this number can be increased up to 200. This may not be sufficient to retrieve the entire information, for example in networks with more than 200 nodes. Not paying attention to paging when using these functions directly (as opposed to within the EcologicalNetworks wrappers) means that you are at risk of not working with the entire dataset.

A note on queries

All queries are passed as vectors of pairs. For example, filtering interactions that are of the mutualist type can be done with [Pair("type", "mutualism")].

For datasets

Mangal.datasets — Function.

datasets()

When called with no arguments, this function will return a list of the most recent datasets. The results will be returned as a vector of MangalDataset object.

source

datasets(q::Vector{Pair{String,T}}) where {T <: Any}

Will return the most recent datasets that match a given query. The results will be returned as a vector of MangalDataset object.

source

Mangal.dataset — Function.

dataset(name::AbstractString)

Return a single dataset by its name.

source

dataset(id::Int64)

Return a single dataset by its unique numerical ID.

source

Networks

networks
network

Interactions

interactions
interaction

Nodes

Mangal.nodes — Function.

nodes()

Gets the latest MangalNode objects.

source

nodes(q::Vector{Pair{String,T}}) where {T <: Any}

Get the latest MangalNode objects according to a given query.

source

nodes(network::MangalNetwork)

Returns the nodes that are part of a MangalNetwork.

source

nodes(network::MangalNetwork, query::Vector{Pair{String,T}}) where {T <: Any}

Returns the nodes that are part of a MangalNetwork, with an additional query.

source

nodes(taxon::MangalReferenceTaxon)

Returns the nodes that are instance of a MangalReferenceTaxon.

source

nodes(taxon::MangalReferenceTaxon, query::Vector{Pair{String,T}}) where {T <: Any}

Returns the nodes that are instance of a MangalReferenceTaxon, with an additional query.

source

Mangal.node — Function.

node(id::Int64)

Returns a node object by id.

source

Reference taxon

Mangal.backbone — Function.

backbone()

Returns the most recent entries in the taxonomic backbone.

source

backbone(q::Vector{Pair{String,T}}) where {T <: Any}

Returns the most recent entries in the taxonomic backbone that match a given query.

source

backbone(name::AbstractString)

Returns the backbone entry for a taxon, matched by exact name.

source

backbone(id::Int64)

Returns the backbone entry for a taxon by id. This function will get the cached version of the backbone if it exists.

source